Remove the "Modeling" nature from the metamodel project.
With the modeling nature, the genmodel is automatically added to the
session but it is not updated by the Ecore modeler when the metamodel
changes. This quickly makes the genmodel out of synch with the
genmodel and causes NPE and other bugs.

Remove the Modeling nature and use the legacy mode to have only the
ecore file in the session.